4.1: moved code to app.py
